Chapter 16: Scanning on Demand
Fixed an issue with `runFile()`'s buffer. Because the buffer had a set size, when the file's length wasn't equivalent to the buffer size, it filled the rest of the array with undefined / ascii 170, which was causing errors in the compiler. Solution was to read the file, get it's byte length, and allocate that amount to the buffer before attempting to read from the file.